+++ src/components/BVN/index.tsx
@@ -3,7 +3,7 @@ import Lady1 from "./assets/Lady with Phone 1.png";
 import Bolt from "./assets/fxemoji_bolt.png";
 import FCMB from "./assets/download.png";
 import "./BVN.css";
-import Frame from "./components/tools/Frame";
+import Frame from "../tools/Frame";
 
 const BVN: React.FC = () => {
   const [selectedOption, setSelectedOption] = useState<"BVN" | "NIN">("BVN");
@@ -14,14 +14,14 @@ const BVN: React.FC = () => {
         <div className="section-1">
           <img src={Lady1} className="img-lady" alt="Lady with Phone" />
           <section>
-          <div className="overlap-group-2">
-                  <p className="everything-is-ea-y">
-                    <span className="span">Ea</span>  <img className="fxemoji-bolt" alt="Fxemoji bolt" src={Bolt} /><span className="span">y loans with </span><span className="FCMB">FastCash</span>
-                   
-                  
-                  </p>
-                 
-                </div> 
+            <div className="overlap-group-2">
+              <p className="everything-is-ea-y">
+                <span className="span">Ea</span>{" "}
+                <img className="fxemoji-bolt" alt="Fxemoji bolt" src={Bolt} />
+                <span className="span">y loans with </span>
+                <span className="FCMB">FastCash</span>
+              </p>
+            </div>
             <div className="frame-3">
               <Frame />
             </div>
@@ -31,7 +31,12 @@ const BVN: React.FC = () => {
           <header>
             <span>
               <p>Go back home </p>
-              <img src={FCMB} className="logo" style={{ width: 50 }} alt="Logo" />
+              <img
+                src={FCMB}
+                className="logo"
+                style={{ width: 50 }}
+                alt="Logo"
+              />
             </span>
           </header>
           <br />
